Mark zeroconf.dns as protected by renaming to zeroconf._dns
- The public API should only access zeroconf and zeroconf.aio
  as internals may be relocated between releases
