+++ zeroconf/core.py
@@ -28,6 +28,7 @@ import threading
 from types import TracebackType  # noqa # used in type hints
 from typing import Dict, List, Optional, Type, Union, cast
 
+from ._dns import DNSIncoming, DNSOutgoing, DNSQuestion
 from .cache import DNSCache
 from .const import (
     _CACHE_CLEANUP_INTERVAL,
@@ -45,7 +46,6 @@ from .const import (
     _TYPE_PTR,
     _UNREGISTER_TIME,
 )
-from .dns import DNSIncoming, DNSOutgoing, DNSQuestion
 from .exceptions import NonUniqueNameException
 from .handlers import QueryHandler, RecordManager
 from .logger import QuietLogger, log
